+++ docs/developer/labs/paulking/sinogram.rst
@@ -1,15 +1,15 @@
 Sinogram
 ========
 
-.. automodule:: pymedphys.sinogram
+.. automodule:: pymedphys._labs.paulking.sinogram
     :no-members:
 
 API
 ---
 
-.. autofunction:: pymedphys.sinogram.read_csv_file
-.. autofunction:: pymedphys.sinogram.read_bin_file
-.. autofunction:: pymedphys.sinogram.crop
-.. autofunction:: pymedphys.sinogram.unshuffle
-.. autofunction:: pymedphys.sinogram.make_histogram
-.. autofunction:: pymedphys.sinogram.find_modulation_factor
+.. autofunction:: pymedphys._labs.paulking.sinogram.read_csv_file
+.. autofunction:: pymedphys._labs.paulking.sinogram.read_bin_file
+.. autofunction:: pymedphys._labs.paulking.sinogram.crop
+.. autofunction:: pymedphys._labs.paulking.sinogram.unshuffle
+.. autofunction:: pymedphys._labs.paulking.sinogram.make_histogram
+.. autofunction:: pymedphys._labs.paulking.sinogram.find_modulation_factor
