+++ src/pymedphys/_labs/paulking/sinogram.csv
