+++ src/pymedphys/_labs/paulking/tomo_sinogram_csv2pdf.py
@@ -36,7 +36,7 @@ from matplotlib.gridspec import GridSpec
 from matplotlib import pyplot as plt
 import numpy as np
 
-from pymedphys.sinogram import unshuffle, crop
+from pymedphys._labs.paulking.sinogram import unshuffle, crop
 
 
 def tomo_sinogram_csv2pdf(file_name='./sinogram.csv', show=True, save=True):
@@ -104,8 +104,8 @@ if __name__ == '__main__':
     import os
 
     try:
-        test = os.path.join(os.getcwd(), 'scripts', 'shareables',
-                            'tomo_sinogram_csv2pdf', 'sinogram.csv')
+        test = os.path.join(os.getcwd(), 'src', 'pymedphys',
+                '_labs', 'paulking', 'sinogram.csv')
         tomo_sinogram_csv2pdf(test, show=True, save=True)
     except IOError:
         print('No sinogram csv file.')
